pyproject.toml
[tool.poetry]
name = "patroni-notifier"
version = "0.0.1"
description = "Patoni notification system using jinja2 templates"
authors = ["Jared Vacanti <jaredvacanti@gmail.com>"]
license = "MIT"
readme = 'README.md'

repository = "https://github.com/jaredvacanti/patroni-notifier"
homepage = "https://github.com/jaredvacanti/patroni-notifier"

keywords = ['patroni', 'patroni-extension', 'ses', 'email', 'patroni-notifier']

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
boto3 = "^1.10.12"
requests = "^2.22.0"
python-consul = "^1.1.0"
jinja2 = "^2.10.3"

[tool.poetry.dev-dependencies]
pytest = "^5.1.3"
pytest-cov = "^2.7.1"
pytest-aiohttp = "^0.3.0"
flake8 = "^3.7.8"
black = "^19.3b0"
pylint = "^2.4.1"
tox = "^3.14"

[tool.poetry.scripts]
swarm-ctl = 'patroni_notifier.core:patroni_notify'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
